+++ units/foundation/include/foundation/allocator.h
@@ -8,8 +8,8 @@
 
 struct allocator_t;
 
-typedef void* (*allocator_alloc_func_t)(allocator_t* allocator, size_t count, size_t size, size_t align, const char* file, int line);
-typedef void* (*allocator_realloc_func_t)(allocator_t* allocator, void* memory, size_t count, size_t size, size_t align, const char* file, int line);
+typedef void* (*allocator_alloc_func_t)(allocator_t* allocator, int64_t count, int64_t size, int64_t align, const char* file, int line);
+typedef void* (*allocator_realloc_func_t)(allocator_t* allocator, void* memory, int64_t count, int64_t size, int64_t align, const char* file, int line);
 typedef void (*allocator_free_func_t)(allocator_t* allocator, void* memory, const char* file, int line);
 
 /**
@@ -22,8 +22,8 @@ struct allocator_t
 	allocator_free_func_t free;
 };
 
-void* allocator_alloc_wrapper(allocator_t* allocator, size_t count, size_t size, size_t align, const char* file, int line);
-void* allocator_realloc_wrapper(allocator_t* allocator, void* memory, size_t count, size_t size, size_t align, const char* file, int line);
+void* allocator_alloc_wrapper(allocator_t* allocator, int64_t count, int64_t size, int64_t align, const char* file, int line);
+void* allocator_realloc_wrapper(allocator_t* allocator, void* memory, int64_t count, int64_t size, int64_t align, const char* file, int line);
 void allocator_free_wrapper(allocator_t* allocator, void* memory, const char* file, int line);
 
 #define ALLOCATOR_ALLOC(allocator, size, align) allocator_alloc_wrapper(allocator, 1, size, align, __FILE__, __LINE__)
@@ -45,12 +45,12 @@ void allocator_free_wrapper(allocator_t* allocator, void* memory, const char* fi
 
 extern allocator_t allocator_malloc;
 
-allocator_t* allocator_incheap_create(allocator_t* parent, size_t size);
+allocator_t* allocator_incheap_create(allocator_t* parent, int64_t size);
 void allocator_incheap_destroy(allocator_t* allocator);
 void allocator_incheap_reset(allocator_t* allocator, void* mark = nullptr);
 void* allocator_incheap_start(allocator_t* allocator);
 void* allocator_incheap_curr(allocator_t* allocator);
-size_t allocator_incheap_bytes_consumed(allocator_t* allocator);
+int64_t allocator_incheap_bytes_consumed(allocator_t* allocator);
 
 struct allocator_helper_t
 {
@@ -64,9 +64,9 @@ struct allocator_helper_t
 allocator_helper_t allocator_helper_create(allocator_t* allocator);
 void* allocator_helper_commit(allocator_helper_t* helper);
 void allocator_helper_destroy(allocator_helper_t* helper);
-void allocator_helper_add(allocator_helper_t* helper, size_t size, size_t align);
-void* allocator_helper_get(allocator_helper_t* helper, size_t size, size_t align);
-size_t allocator_helper_size(allocator_helper_t* helper);
+void allocator_helper_add(allocator_helper_t* helper, int64_t size, int64_t align);
+void* allocator_helper_get(allocator_helper_t* helper, int64_t size, int64_t align);
+int64_t allocator_helper_size(allocator_helper_t* helper);
 
 #define ALLOCATOR_HELPER_ADD(helper, n, type) allocator_helper_add(helper, n*sizeof(type), ALIGNOF(type))
 #define ALLOCATOR_HELPER_GET(helper, n, type) (type*)allocator_helper_get(helper, n*sizeof(type), ALIGNOF(type))
